refactor(#1143): split scan_stream_windowed tests to sibling file (size ratchet)
The batching change pushed scan_stream_windowed.rs to 823 lines, over the 800-line
source limit, so the agent-gate file-size ratchet failed (campsite rule). Extract
the inline #[cfg(test)] mod tests into scan_stream_windowed_tests.rs, included via
#[cfg(test)] #[path=...] mod tests; so use super::* still resolves to the
windowed-scan module's private items. Parent now 616 lines, test sibling 217. No
behavior change; both the pure terminal-drain guard and the dataset-dependent
fixture guard still run, and the scan-offload guard still passes.
